Created the Dircomponent component
The Dircomponent mimics the folder structure of the UI located at the
left.
This is implemented through nesting the component in itself.
The compoenent is rendered as a list <li> element so its parent must
call a <ul> tag and encapsulate this dircomponent.
This component receives a prop named data from the main App.js
data must contain the key value pairs of 'title' and 'children'
'title' - represents the name of the folder or file.
'children' - is an array that contains none or more objects similar to this
data object.
If this component contains children it can be expanded on click to
render dircomponents with the value of each 'children' passed as the
data prop.

State is managed within the component and not at main

**Further updates will require icon representation of each list element.