Child menu items are now added
When hovering over menu items a child menu pops up on the right side of
the menu. These Menu items can have children as well.

This is achieved by adding the Menuitem.js component. The menu items are
stored in the main App.js as part of the menu state. In the state
variable for the menu, when a menu object has the children object in it
it is expandable.

** Possible additions would be the arrow symbol to indicate that a menu
item has children, and the _________ for separating related menu items.